pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "platform-challenge-sdk"
version = "0.1.0"
description = "Python SDK for developing challenges on Platform Network"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Platform", email = "dev@platform.network" }]
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "requests>=2.32.0",
    "pynacl>=1.5.0",
    "cryptography>=42.0.0",
    "pycryptodome>=3.20.0",
    "dstack-sdk>=0.5.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy[asyncio]>=2.0.23",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "black>=24.8.0",
    "isort>=5.13.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[tool.hatchling.build.targets.wheel]
packages = ["src/platform_challenge_sdk"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "UP", "B", "S", "ASYNC", "A", "C4", "PIE", "Q"]
format.preview = true
fix = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

